move news components
